+++ app/src/main/java/com/growthook/aos/presentation/cavecreate/CreateNewCaveActivity.kt
@@ -5,7 +5,7 @@ import android.os.Bundle
 import android.view.MotionEvent
 import android.widget.Toast
 import androidx.activity.viewModels
-import com.growthook.aos.presentation.cavecreate.model.NewCaveIntent
+import com.growthook.aos.presentation.model.NewCaveIntent
 import com.growthook.aos.databinding.ActivityCreateNewCaveBinding
 import com.growthook.aos.util.base.BaseActivity
 import com.growthook.aos.util.base.BaseAlertDialog
